bug 8561 fixed: DIV function not working
git-svn-id: http://talendforge.org/svn/tos/trunk@27321 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.librariesmanager/resources/java/routines/Mathematical.java
@@ -198,8 +198,8 @@ public class Mathematical {
      * {example} DIV(3.14,3.14)
      * 
      */
-    public static double DIV(double a, double b) {
-        return a / b;
+    public static int DIV(double a, double b) {
+        return (int) (a / b);
     }
 
     /**
